Cleaning up a bit for the scraper
