+++ src/go-apps/flame-scrape/go.mod
@@ -1,11 +1,8 @@
-module github.com/TrueBlocks/trueblocks-core/src/go-apps/blaze
+module github.com/TrueBlocks/trueblocks-core/src/go-apps/flame-scrape
 
 go 1.15
 
 require (
-	github.com/gorilla/mux v1.8.0
 	github.com/gorilla/websocket v1.4.2
-	github.com/mitchellh/go-homedir v1.1.0
 	github.com/spf13/pflag v1.0.5
-	golang.org/x/time v0.0.0-20210220033141-f8bda1e9f3ba
 )
