+++ src/go-apps/flame-scrape/scraper/scraper_index.go
@@ -1,4 +1,4 @@
-package server
+package scraper
 
 /*-------------------------------------------------------------------------------------------
  * qblocks - fast, easily-accessible, fully-decentralized data from blockchains
@@ -20,7 +20,7 @@ import (
 	"strconv"
 	"sync"
 
-	utils "github.com/TrueBlocks/trueblocks-core/src/go-apps/blaze/utils"
+	utils "github.com/TrueBlocks/trueblocks-core/src/go-apps/flame-scrape/utils"
 )
 
 var IndexScraper Scraper
