--- src/go-apps/flame-scrape/server/server_progress.go
@@ -1,60 +1,0 @@
-package server
-
-/*-------------------------------------------------------------------------------------------
- * qblocks - fast, easily-accessible, fully-decentralized data from blockchains
- * copyright (c) 2016, 2021 TrueBlocks, LLC (http://trueblocks.io)
- *
- * This program is free software: you may redistribute it and/or modify it under the terms
- * of the GNU General Public License as published by the Free Software Foundation, either
- * version 3 of the License, or (at your option) any later version. This program is
- * distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * General Public License for more details. You should have received a copy of the GNU General
- * Public License along with this program. If not, see http://www.gnu.org/licenses/.
- *-------------------------------------------------------------------------------------------*/
-
-import (
-	"bufio"
-	"bytes"
-	"fmt"
-	"io"
-	"strings"
-)
-
-// dropNL drops new line characters (\n) from the progress stream
-func dropNL(data []byte) []byte {
-	if len(data) > 0 && data[len(data)-1] == '\n' {
-		return data[0 : len(data)-1]
-	}
-	return data
-}
-
-// ScanProgressLine looks for "lines" that end with `\r` not `\n` like usual
-func ScanProgressLine(data []byte, atEOF bool) (advance int, token []byte, err error) {
-	if atEOF && len(data) == 0 {
-		return 0, nil, nil
-	}
-	if i := bytes.IndexByte(data, '\r'); i >= 0 {
-		return i + 1, dropNL(data[0:i]), nil
-	}
-	return bufio.ScanLines(data, atEOF)
-}
-
-// ScanForProgress watches stderr and picks of progress messages
-func ScanForProgress(stderrPipe io.Reader, fn func(string)) {
-	scanner := bufio.NewScanner(stderrPipe)
-	scanner.Split(ScanProgressLine)
-	for scanner.Scan() {
-		text := scanner.Text()
-		if len(text) > 0 {
-			fmt.Println(text)
-			if strings.Contains(text, "<PROG>") {
-				fn(strings.SplitAfter(text, ":")[1])
-			}
-		}
-	}
-
-	if err := scanner.Err(); err != nil {
-		fmt.Println("Error while reading stderr:", err)
-	}
-}
